+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# unreleased
+
+* Add both with- and without-message variants for bulk unsubscribe test helpers (for Email Alert API)
+* BREAKING: Add `sender_message_id` and `govuk_request_id` to bulk unsubscribe (bad request) test helper (for Email Alert API)
+* Add `content_id` to subscriber lists URL helper (for Email Alert API)
+
 # 78.1.0
 
 * Fix `bulk_unsubscribe` requires a `govuk_request_id` if you want to send a message